Fix tty detection

+++ .bashrc.d/020aliases.bashrc
@@ -215,7 +215,7 @@ function rclone_move_callisto () {
 }
 
 function rclone () {
-    if [[ -t 1 ]] &&  [[ -t 2 ]]
+    if [[ -t 1 ]] &&  [[ -t 2 ]] && [[ ! -p /dev/stdout ]] && [[ ! -p /dev/stdin ]]
     then
         echo 'Detected TTY' >&2
         docker run -it --rm --net host -v /root:/root -v "$(pwd):$(pwd)" -w "$(pwd)" --entrypoint rclone peelvalley/rclone-b2 "${@}"
